修复 stat fstat 系统调用错误返回值, 修复 read write 没有文件读取偏移问题, 修复文件错误 inode 计数问题(临时解决, 可运行 rm ls 命令)

+++ src/x86_64/include/vdisk.h
@@ -22,8 +22,8 @@ typedef struct {
     char         drive_name[50];
 } vdisk; // 块设备
 
-uint32_t disk_size(int drive);
-int      vdisk_init();
+size_t disk_size(int drive);
+int    vdisk_init();
 
 /**
  * 注册一个硬盘设备
